hookjavac

+++ DevSample/buildSrc/src/main/groovy/com/dx168/fastdex/build/task/FastdexResourceIdTask.groovy
@@ -1,5 +1,6 @@
 package com.dx168.fastdex.build.task
 
+import com.dx168.fastdex.build.util.Constant
 import com.dx168.fastdex.build.util.FileUtils
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
@@ -7,6 +8,7 @@ import com.tencent.tinker.build.aapt.AaptResourceCollector
 import com.tencent.tinker.build.aapt.AaptUtil
 import com.tencent.tinker.build.aapt.PatchUtil
 import com.tencent.tinker.build.aapt.RDotTxtEntry
+import com.dx168.fastdex.build.util.FastdexUtils
 
 /**
  * Created by tong on 17/3/11.
@@ -24,14 +26,23 @@ public class FastdexResourceIdTask extends DefaultTask {
 
     @TaskAction
     def applyResourceId() {
-        File buildDir = FileUtils.getFastdexBuildDir(project,variantName)
-        String resourceMappingFile = new File(buildDir,FileUtils.R_TXT)
+        File buildDir = FastdexUtils.getBuildDir(project,variantName)
+
+        String resourceMappingFile = new File(buildDir,Constant.R_TXT)
 
         // Parse the public.xml and ids.xml
         if (!FileUtils.isLegalFile(resourceMappingFile)) {
-            project.logger.error("apply resource mapping file ${resourceMappingFile} is illegal, just ignore")
+            project.logger.error("==fastdex apply resource mapping file ${resourceMappingFile} is illegal, just ignore")
             return
         }
+
+        File idsXmlFile = new File(buildDir,RESOURCE_IDX_XML)
+        File publicXmlFile = new File(buildDir,RESOURCE_PUBLIC_XML)
+        if (FileUtils.isLegalFile(idsXmlFile) && FileUtils.isLegalFile(publicXmlFile)) {
+            project.logger.error("==fastdex public xml file and ids xml file already exist, just ignore")
+            return
+        }
+
         String idsXml = resDir + "/values/ids.xml";
         String publicXml = resDir + "/values/public.xml";
         FileUtils.deleteFile(idsXml);
@@ -39,7 +50,7 @@ public class FastdexResourceIdTask extends DefaultTask {
         List<String> resourceDirectoryList = new ArrayList<String>()
         resourceDirectoryList.add(resDir)
 
-        project.logger.error("we build ${project.getName()} apk with apply resource mapping file ${resourceMappingFile}")
+        project.logger.error("==fastdex we build ${project.getName()} apk with apply resource mapping file ${resourceMappingFile}")
         Map<RDotTxtEntry.RType, Set<RDotTxtEntry>> rTypeResourceMap = PatchUtil.readRTxt(resourceMappingFile)
 
         AaptResourceCollector aaptResourceCollector = AaptUtil.collectResource(resourceDirectoryList, rTypeResourceMap)
@@ -48,12 +59,12 @@ public class FastdexResourceIdTask extends DefaultTask {
 
 
         if (publicFile.exists()) {
-            FileUtils.copyFileUsingStream(publicFile, new File(buildDir,RESOURCE_PUBLIC_XML))
+            FileUtils.copyFileUsingStream(publicFile, publicXmlFile)
             project.logger.error("==fastdex gen resource public.xml in ${RESOURCE_PUBLIC_XML}")
         }
         File idxFile = new File(idsXml)
         if (idxFile.exists()) {
-            FileUtils.copyFileUsingStream(idxFile, new File(buildDir,RESOURCE_IDX_XML))
+            FileUtils.copyFileUsingStream(idxFile, idsXmlFile)
             project.logger.error("==fastdex gen resource idx.xml in ${RESOURCE_IDX_XML}")
         }
     }
